All tests passing with 100% coverage
