+++ app/controllers/welcome_page_controller.rb
@@ -5,12 +5,8 @@ class WelcomePageController < ApplicationController
   def sign_in; end
 
   def timeline
-    if user_signed_in?
-      @posts = Post.all
-      @image_posts = ImagePost.all
-      @user_email = current_user.email
-    else
-      render :landing_page
-    end
+    @posts = Post.all
+    @image_posts = ImagePost.all
+    @user_email = current_user.email
   end
 end
